+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file.
 - (hooks) Added `useTheme()` hook
 - (hooks) Added `useScript()` hook
 - (hooks) Added `useVisibility()` hook
+- (hooks) Added `useOutsideClick()` hook
 
 ### Changed
 - (config) Moved typescript-config into a config package (@enotion/config/typescript)